feat: implement flying navigation, shoulder riding and the parrot
The parrot needed two foundations Steel did not have, so they land here rather
than inside the mob.

Flying navigation: `FlyNodeEvaluator` joins the walking and swimming ones
behind the existing `NodeEvaluator` trait, `NavigationKind::Flying` selects it,
and the navigation tick feeds the path position straight to the move control
instead of dropping it to the floor. `FlyingMoveControl` arrives as
`MoveControlKind`, mirroring how `NavigationKind` already lets a mob choose how
it finds its way. `WaterAvoidingRandomFlyingGoal` and the `HoverRandomPos` /
`AirAndWaterRandomPos` helpers come with it.

Shoulder riding: a player now carries up to two entities, drops them when it
falls, swims, flies, sleeps, stands in powder snow, dies or turns spectator,
and lets a parrot on one speak for it. `LandOnOwnersShoulderGoal` is what puts
the bird there.

The parrot itself: seed taming, sitting, the five variants, the cookie that
poisons and kills it, and the forty-entry imitation table that makes it worth
carrying at all.

Two gaps are named in the source. A shoulder rider is returned to the world
when its owner logs out rather than saved with the player, because
`PersistentPlayerData` has no field for one yet. And `Parrot.doPush`, which
stops a parrot shoving players, has no per-pusher hook to override: Steel
resolves pushing on the pushed entity.

+++ steel-core/src/entity/ai/goal/ocelot_attack.rs
@@ -92,7 +92,7 @@ impl Goal for OcelotAttackGoal {
             30.0,
         );
 
-        let melee_radius = f64::from(mob.bounding_box().width() * 2.0);
+        let melee_radius = mob.bounding_box().width() * 2.0;
         let melee_radius_sqr = melee_radius * melee_radius;
         let distance_sqr = mob.position().distance_squared(target_position);
         let speed_modifier =
